+++ ff.h
@@ -116,11 +116,14 @@ Boolean setup(FilterRecordPtr pb);
 void evalpixel(unsigned char *outp,unsigned char *inp);
 
 // from make.c
-extern const volatile uint32_t cObfuscV4Seed; // this value will be manipulated during the building of each individual filter (see make_win.c)
 unsigned long printablehash(unsigned long hash);
 size_t fixpipl(PIPropertyList *pipl,size_t origsize,StringPtr title, long *event_id);
 size_t aete_generate(void* aeteptr, PARM_T *pparm, long event_id);
-void obfusc(PARM_T* pparm, unsigned int seed);
+
+// from obfusc.c
+extern const volatile uint32_t cObfuscV4Seed; // this value will be manipulated during the building of each individual filter (see make_win.c)
+int obfuscationVersion(PARM_T* pparm);
+unsigned int obfusc(PARM_T* pparm);
 void deobfusc(PARM_T* pparm);
 
 // from loadfile_*.c