+++ load_win.c
@@ -91,16 +91,12 @@ Boolean loadfile(StandardFileReply *sfr,char **reason){
 		if (hm = LoadLibraryEx(myp2cstrcpy(name,sfr->sfFile.name),NULL,LOAD_LIBRARY_AS_DATAFILE)) {
 			if (readPARMresource(hm,reason,READ_OBFUSC)) {
 				if ((gdata->parm.cbSize != PARM_SIZE) && (gdata->parm.cbSize != PARM_SIZE_PREMIERE) && (gdata->parm.cbSize != PARM_SIG_MAC)) {
-					if (gdata->parm.unknown2 == 4) {
-						// Obfuscation V4 is protected, because FF>=1.7.0.5 combines protection and obfuscation
-						*reason = _strdup("The filter is protected.");
-					}
-					else {
-						*reason = _strdup("Incompatible obfuscation.");
-					}
+					*reason = _strdup("Incompatible obfuscation.");
+					//gdata->parmloaded = false;
 					return false; // Stop! We know the issue now.
 				} else if (gdata->parm.iProtected) {
 					*reason = _strdup("The filter is protected.");
+					//gdata->parmloaded = false;
 					return false; // Stop! We know the issue now.
 				} else {
 					readok = gdata->parmloaded = true;